+++ tests/integration/test_target_snowflake.py
@@ -3,6 +3,7 @@ import json
 import unittest
 import mock
 import os
+import botocore
 
 from nose.tools import assert_raises
 
@@ -37,6 +38,10 @@ class TestIntegration(unittest.TestCase):
         # Drop target schema
         if self.config['default_target_schema']:
             snowflake.query("DROP SCHEMA IF EXISTS {}".format(self.config['default_target_schema']))
+
+    def persist_lines(self, lines):
+        """Loads singer messages into snowflake without table caching option"""
+        target_snowflake.persist_lines(self.config, lines)
 
     def persist_lines_with_cache(self, lines):
         """Enables table caching option and loads singer messages into snowflake.
@@ -571,27 +576,92 @@ class TestIntegration(unittest.TestCase):
             table_one,
             [{'C_INT': 1, 'C_PK': 1, 'C_VARCHAR': '1'}])
 
-        # Table two should have versioned column
+        # Table two should have a versioned column and a new column
         self.assertEquals(
             table_two,
             [
                 {previous_column_name: datetime.datetime(2019, 2, 1, 15, 12, 45), 'C_INT': 1, 'C_PK': 1,
-                 'C_VARCHAR': '1', 'C_DATE': None},
+                 'C_VARCHAR': '1', 'C_DATE': None, 'C_NEW_COLUMN': None},
                 {previous_column_name: datetime.datetime(2019, 2, 10, 2), 'C_INT': 2, 'C_PK': 2, 'C_VARCHAR': '2',
-                 'C_DATE': '2019-02-12 02:00:00'},
-                {previous_column_name: None, 'C_INT': 3, 'C_PK': 3, 'C_VARCHAR': '2', 'C_DATE': '2019-02-15 02:00:00'}
+                 'C_DATE': '2019-02-12 02:00:00', 'C_NEW_COLUMN': 'data 1'},
+                {previous_column_name: None, 'C_INT': 3, 'C_PK': 3, 'C_VARCHAR': '2', 'C_DATE': '2019-02-15 02:00:00',
+                 'C_NEW_COLUMN': 'data 2'}
             ]
         )
 
-        # Table three should have renamed columns
+        # Table three should have a renamed columns and a new column
         self.assertEqual(
             table_three,
             [
-                {'C_INT': 1, 'C_PK': 1, 'C_TIME': datetime.time(4, 0), 'C_VARCHAR': '1', 'C_TIME_RENAMED': None},
-                {'C_INT': 2, 'C_PK': 2, 'C_TIME': datetime.time(7, 15), 'C_VARCHAR': '2', 'C_TIME_RENAMED': None},
+                {'C_INT': 1, 'C_PK': 1, 'C_TIME': datetime.time(4, 0), 'C_VARCHAR': '1', 'C_TIME_RENAMED': None,
+                 'C_NEW_COLUMN': None},
+                {'C_INT': 2, 'C_PK': 2, 'C_TIME': datetime.time(7, 15), 'C_VARCHAR': '2', 'C_TIME_RENAMED': None,
+                 'C_NEW_COLUMN': None},
                 {'C_INT': 3, 'C_PK': 3, 'C_TIME': datetime.time(23, 0, 3), 'C_VARCHAR': '3',
-                 'C_TIME_RENAMED': datetime.time(8, 15)},
-                {'C_INT': 4, 'C_PK': 4, 'C_TIME': None, 'C_VARCHAR': '4', 'C_TIME_RENAMED': datetime.time(23, 0, 3)}
+                 'C_TIME_RENAMED': datetime.time(8, 15), 'C_NEW_COLUMN': 'data 1'},
+                {'C_INT': 4, 'C_PK': 4, 'C_TIME': None, 'C_VARCHAR': '4', 'C_TIME_RENAMED': datetime.time(23, 0, 3),
+                 'C_NEW_COLUMN': 'data 2'}
+            ])
+
+    def test_column_name_change_without_table_cache(self):
+        """Tests correct renaming of snowflake columns after source change with not using table caching"""
+        tap_lines_before_column_name_change = test_utils.get_test_tap_lines('messages-with-three-streams.json')
+        tap_lines_after_column_name_change = test_utils.get_test_tap_lines(
+            'messages-with-three-streams-modified-column.json')
+
+        # Load with default settings
+        self.persist_lines(tap_lines_before_column_name_change)
+        self.persist_lines(tap_lines_after_column_name_change)
+
+        # Get loaded rows from tables
+        snowflake = DbSync(self.config)
+        target_schema = self.config.get('default_target_schema', '')
+        table_one = snowflake.query("SELECT * FROM {}.test_table_one ORDER BY c_pk".format(target_schema))
+        table_two = snowflake.query("SELECT * FROM {}.test_table_two ORDER BY c_pk".format(target_schema))
+        table_three = snowflake.query("SELECT * FROM {}.test_table_three ORDER BY c_pk".format(target_schema))
+
+        # Get the previous column name from information schema in test_table_two
+        previous_column_name = snowflake.query("""
+            SELECT column_name
+              FROM information_schema.columns
+             WHERE table_catalog = '{}'
+               AND table_schema = '{}'
+               AND table_name = 'TEST_TABLE_TWO'
+               AND ordinal_position = 1
+            """.format(
+            self.config.get('dbname', '').upper(),
+            target_schema.upper()))[0]["COLUMN_NAME"]
+
+        # Table one should have no changes
+        self.assertEqual(
+            table_one,
+            [{'C_INT': 1, 'C_PK': 1, 'C_VARCHAR': '1'}])
+
+        # Table two should have a versioned column and a new column
+        self.assertEquals(
+            table_two,
+            [
+                {previous_column_name: datetime.datetime(2019, 2, 1, 15, 12, 45), 'C_INT': 1, 'C_PK': 1,
+                 'C_VARCHAR': '1', 'C_DATE': None, 'C_NEW_COLUMN': None},
+                {previous_column_name: datetime.datetime(2019, 2, 10, 2), 'C_INT': 2, 'C_PK': 2, 'C_VARCHAR': '2',
+                 'C_DATE': '2019-02-12 02:00:00', 'C_NEW_COLUMN': 'data 1'},
+                {previous_column_name: None, 'C_INT': 3, 'C_PK': 3, 'C_VARCHAR': '2', 'C_DATE': '2019-02-15 02:00:00',
+                 'C_NEW_COLUMN': 'data 2'}
+            ]
+        )
+
+        # Table three should have a renamed columns and a new column
+        self.assertEqual(
+            table_three,
+            [
+                {'C_INT': 1, 'C_PK': 1, 'C_TIME': datetime.time(4, 0), 'C_VARCHAR': '1', 'C_TIME_RENAMED': None,
+                 'C_NEW_COLUMN': None},
+                {'C_INT': 2, 'C_PK': 2, 'C_TIME': datetime.time(7, 15), 'C_VARCHAR': '2', 'C_TIME_RENAMED': None,
+                 'C_NEW_COLUMN': None},
+                {'C_INT': 3, 'C_PK': 3, 'C_TIME': datetime.time(23, 0, 3), 'C_VARCHAR': '3',
+                 'C_TIME_RENAMED': datetime.time(8, 15), 'C_NEW_COLUMN': 'data 1'},
+                {'C_INT': 4, 'C_PK': 4, 'C_TIME': None, 'C_VARCHAR': '4', 'C_TIME_RENAMED': datetime.time(23, 0, 3),
+                 'C_NEW_COLUMN': 'data 2'}
             ])
 
     def test_logical_streams_from_pg_with_hard_delete_and_default_batch_size_should_pass(self):
@@ -846,24 +916,89 @@ class TestIntegration(unittest.TestCase):
 
         self.assert_three_streams_are_into_snowflake()
 
-    def test_using_aws_environment_variables(self):
-        """Test loading data with aws in the environment rather than explicitly provided access keys"""
+    def test_aws_env_vars(self):
+        """Test loading data with credentials defined in AWS environment variables
+        than explicitly provided access keys"""
         tap_lines = test_utils.get_test_tap_lines("messages-with-three-streams.json")
 
         try:
-            os.environ["AWS_ACCESS_KEY_ID"] = os.environ.get(
-                "TARGET_SNOWFLAKE_AWS_ACCESS_KEY"
-            )
-            os.environ["AWS_SECRET_ACCESS_KEY"] = os.environ.get(
-                "TARGET_SNOWFLAKE_AWS_SECRET_ACCESS_KEY"
-            )
-            self.config["aws_access_key_id"] = None
-            self.config["aws_secret_access_key"] = None
+            # Save original config to restore later
+            orig_config = self.config.copy()
 
-            target_snowflake.persist_lines(self.config, tap_lines)
+            # Move aws access key and secret from config into environment variables
+            os.environ['AWS_ACCESS_KEY_ID'] = os.environ.get('TARGET_SNOWFLAKE_AWS_ACCESS_KEY')
+            os.environ['AWS_SECRET_ACCESS_KEY'] = os.environ.get('TARGET_SNOWFLAKE_AWS_SECRET_ACCESS_KEY')
+            del self.config['aws_access_key_id']
+            del self.config['aws_secret_access_key']
+
+            # Create a new S3 client using env vars
+            snowflake = DbSync(self.config)
+            snowflake.create_s3_client()
+
+        # Restore the original state to not confuse other tests
         finally:
-            del os.environ["AWS_ACCESS_KEY_ID"]
-            del os.environ["AWS_SECRET_ACCESS_KEY"]
+            del os.environ['AWS_ACCESS_KEY_ID']
+            del os.environ['AWS_SECRET_ACCESS_KEY']
+            self.config = orig_config.copy()
+
+    def test_profile_based_auth(self):
+        """Test AWS profile based authentication rather than access keys"""
+        try:
+            # Save original config to restore later
+            orig_config = self.config.copy()
+
+            # Remove access keys from config and add profile name
+            del self.config['aws_access_key_id']
+            del self.config['aws_secret_access_key']
+            self.config['aws_profile'] = 'fake-profile'
+
+            # Create a new S3 client using profile based authentication
+            with assert_raises(botocore.exceptions.ProfileNotFound):
+                snowflake = DbSync(self.config)
+                snowflake.create_s3_client()
+
+        # Restore the original state to not confuse other tests
+        finally:
+            self.config = orig_config.copy()
+
+    def test_profile_based_auth_aws_env_var(self):
+        """Test AWS profile based authentication using AWS environment variables"""
+        try:
+            # Save original config to restore later
+            orig_config = self.config.copy()
+
+            # Remove access keys from config and add profile name environment variable
+            del self.config['aws_access_key_id']
+            del self.config['aws_secret_access_key']
+            os.environ['AWS_PROFILE'] = 'fake_profile'
+
+            # Create a new S3 client using profile based authentication
+            with assert_raises(botocore.exceptions.ProfileNotFound):
+                snowflake = DbSync(self.config)
+                snowflake.create_s3_client()
+
+        # Restore the original state to not confuse other tests
+        finally:
+            del os.environ['AWS_PROFILE']
+            self.config = orig_config.copy()
+
+    def test_s3_custom_endpoint_url(self):
+        """Test S3 connection with custom region and endpoint URL"""
+        try:
+            # Save original config to restore later
+            orig_config = self.config.copy()
+
+            # Define custom S3 endpoint
+            self.config['s3_endpoint_url'] = 'fake-endpoint-url'
+
+            # Botocore should raise ValurError in case of fake S3 endpoint url
+            with assert_raises(ValueError):
+                snowflake = DbSync(self.config)
+                snowflake.create_s3_client()
+
+        # Restore the original state to not confuse other tests
+        finally:
+            self.config = orig_config.copy()
 
     def test_too_many_records_exception(self):
         """Test if query function raise exception if max_records exceeded"""
@@ -886,3 +1021,22 @@ class TestIntegration(unittest.TestCase):
         self.persist_lines_with_cache(tap_lines)
 
         self.assert_three_streams_are_into_snowflake()
+
+    def test_quoted_identifiers_ignore_case_session_parameter(self):
+        """Test if QUOTED_IDENTIFIERS_IGNORE_CASE session parameter set to FALSE"""
+        snowflake = DbSync(self.config)
+
+        # Set QUOTED_IDENTIFIERS_IGNORE_CASE to True on user level
+        snowflake.query(f"ALTER USER {self.config['user']} SET QUOTED_IDENTIFIERS_IGNORE_CASE = TRUE")
+
+        # Quoted column names should be case sensitive even if the
+        # QUOTED_IDENTIFIERS_IGNORE_CASE parameter set to TRUE on user or account level
+        result = snowflake.query('SELECT 1 AS "Foo", 1 AS "foo", 1 AS "FOO", 1 AS foo, 1 AS FOO')
+        self.assertEqual(result, [{
+            'Foo': 1,
+            'foo': 1,
+            'FOO': 1
+        }])
+
+        # Reset parameters default
+        snowflake.query(f"ALTER USER {self.config['user']} UNSET QUOTED_IDENTIFIERS_IGNORE_CASE")