+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+1.8.0 (2020-08-03)
+-------------------
+
+- Fixed an issue when `pipelinewise-target-snowflake` failed when `QUOTED_IDENTIFIERS_IGNORE_CASE` snowflake parameter set to true
+- Add `aws_profile` option to support Profile based authentication to S3
+- Add option to authenticate to S3 using `AWS_PROFILE`, `AWS_ACCESS_KEY_ID`, `AWS_SECRET_ACCESS_KEY` and `AWS_SESSION_TOKEN` environment variables
+- Add `s3_endpoint_url` and `s3_region_name` options to support non-native S3 accounts
+- Flush stream only if the new schema is not the same as the previous one
+
+1.7.0 (2020-07-23)
+-------------------
+
+- Add `s3_acl` option to support ACL for S3 upload
+- Fixed an issue when no primary key error logged as `INFO` and not as `ERROR`
+
+1.6.6 (2020-06-26)
+-------------------
+
+- Fixed an issue when new columns sometimes not added to target table
+- Fixed an issue when the query runner returned incorrect value when multiple queries running in one transaction
+
 1.6.5 (2020-06-17)
 -------------------